Move App class out of global namespace
